Don't call pthread_atfork from a the atfork handler
brpc is creating a metrics sampler thread at load time. It also tries
to make sure that, if fork() is called, the child process recreates
this sampler thread, since the child process will only have a clone of
the thread calling fork() from the parent.

It does this by calling pthread_atfork() to register a callback that
will be invoked in the child process. This callback recreates the
sampler thread in the child.

I've repeatedly seen a process get stuck in this atfork handler
because it is calling pthread_atfork() to register the callback again
in the child. There are two problems with this.

1. pthread_atfork is not considered reentrant within the context of an
   atfork handler. Indeed, on the system I'm using (Ubuntu 20.04) this
   leads to deadlock since the caller of the atfork handler is holding
   a lock required by pthread_atfork().
2. These handlers are inherited by the child process so there is no
   need to re-register them.

This just moves the call to pthread_atfork() to the SamplerCollector
constructor where it will only be called once.